reorganize sorted fs changes
